requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymupdf==1.23.8
python-multipart==0.0.6
Pillow==10.1.0
requests==2.31.0 